requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
asgiref==3.4.1
    # via django
bleach==4.1.0
    # via django-markdownify
boto3==1.17.35
    # via -r requirements.in
botocore==1.20.112
    # via
    #   boto3
    #   s3transfer
certifi==2021.10.8
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==2.0.7
    # via requests
cryptography==35.0.0
    # via social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
discord-interactions==0.3.0
    # via -r requirements.in
dj-database-url==0.5.0
    # via -r requirements.in
django==3.1.6
    # via
    #   -r requirements.in
    #   django-ace
    #   django-anymail
    #   django-filter
    #   django-markdownify
    #   django-storages
    #   djangorestframework
    #   stream-django
django-ace==1.0.11
    # via -r requirements.in
django-anymail[mailgun]==8.2
    # via -r requirements.in
django-email-verification==0.1.0
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
django-markdownify==0.9.0
    # via -r requirements.in
django-mdeditor==0.1.18
    # via -r requirements.in
django-storages==1.11.1
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
gunicorn==20.0.4
    # via -r requirements.in
idna==3.3
    # via requests
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
markdown==3.3.4
    # via
    #   -r requirements.in
    #   django-markdownify
midtransclient==1.2.0
    # via -r requirements.in
oauthlib==3.1.1
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==21.0
    # via bleach
pillow==8.1.0
    # via -r requirements.in
psycopg2-binary==2.8.6
    # via -r requirements.in
pycparser==2.20
    # via cffi
pycryptodomex==3.11.0
    # via stream-python
pyjwt==2.3.0
    # via
    #   social-auth-core
    #   stream-python
pynacl==1.4.0
    # via discord-interactions
pyparsing==2.4.7
    # via packaging
python-dateutil==2.8.2
    # via botocore
python3-openid==3.2.0
    # via social-auth-core
pytz==2021.3
    # via
    #   django
    #   djangorestframework
    #   stream-django
    #   stream-python
pyyaml==6.0
    # via -r requirements.in
requests==2.26.0
    # via
    #   django-anymail
    #   midtransclient
    #   requests-oauthlib
    #   social-auth-core
    #   stream-python
requests-oauthlib==1.3.0
    # via social-auth-core
s3transfer==0.3.7
    # via boto3
sentry-sdk==1.4.3
    # via -r requirements.in
six==1.16.0
    # via
    #   bleach
    #   pynacl
    #   python-dateutil
    #   social-auth-app-django
social-auth-app-django==4.0.0
    # via -r requirements.in
social-auth-core==4.1.0
    # via social-auth-app-django
sorl-thumbnail==12.7.0
    # via -r requirements.in
sqlparse==0.4.2
    # via django
stream-django==1.7.1
    # via -r requirements.in
stream-python==5.1.0
    # via stream-django
urllib3==1.26.7
    # via
    #   botocore
    #   requests
    #   sentry-sdk
webencodings==0.5.1
    # via bleach
whitenoise==5.2.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
